Refactored DAO so storage specific code is in a Datasource. DAO now deals with navigation only.
