+++ src/DAO/Dao.ts
@@ -1,12 +1,21 @@
+import DaoRecord from "./DaoRecord";
 import IDao from "./IDao";
 import IDaoFactory from "./IDaoFactory";
 import IDaoFieldDef from "./IDaoFieldDef";
 import IDaoRecord from "./IDaoRecord";
+import IDatasource from "./IDatasource";
+import { DaoType } from "./types";
 
 export default abstract class Dao implements IDao {
-    protected readonly _factory: IDaoFactory;
+    private readonly _factory: IDaoFactory;
+    private readonly _datasource: IDatasource = null;
 
-    protected _records: Array<IDaoRecord> = new Array();
+    readonly objectName: string;
+    get fieldCount(): number {
+        return this._datasource.fieldDefs.length;
+    }
+
+    private _records: Array<IDaoRecord> = new Array();
     get recordCount(): number {
         return this._records.length;
     }
@@ -17,48 +26,149 @@ export default abstract class Dao implements IDao {
         return this._records[x];
     }
 
-    protected _fieldDefs: Array<IDaoFieldDef> = new Array();
-    get fieldCount(): number {
-        return this._fieldDefs.length;
-    }
-    getFieldDef(x: number): IDaoFieldDef {
-        if (x < 0 || x >= this._fieldDefs.length) {
-            return null;
-        }
-        return this._fieldDefs[x];
-    }
-
-    protected _currentRecordIndex: number = null;
-    protected _currentRecord: IDaoRecord = null;
+    private _currentRecordIndex: number = null;
+    private _currentRecord: IDaoRecord = null;
     get currentRecord(): IDaoRecord {
         return this._currentRecord;
     }
 
-    protected _bof: boolean = true;
+    private _bof: boolean = true;
     get bof(): boolean {
         return this._bof;
     }
 
-    protected _eof: boolean = false;
+    private _eof: boolean = false;
     get eof(): boolean {
         return this._eof;
     }
 
-    readonly objectName: string = null;
 
-
-    constructor(daoFactory: IDaoFactory, objectName: string) {
+    constructor(daoFactory: IDaoFactory, datasource: IDatasource) {
         this._factory = daoFactory;
-        this.objectName = objectName;
+        this._datasource = datasource;
     }
 
-    abstract load(fields?: Array<string>, filter?: string, maxRows?: number): Promise<number>;
-    abstract save(): Promise<number>;
-    abstract discard(): number;
-    abstract addRecord(): IDaoRecord;
-    abstract addRecord(values?: Map<string, any>): IDaoRecord;
-    abstract first(): IDaoRecord;
-    abstract last(): Promise<IDaoRecord>;
-    abstract prev(): IDaoRecord;
-    abstract next(): Promise<IDaoRecord>;
+    async load(fields?: Array<string>, filter?: string, maxRows?: number): Promise<number> {
+        this._currentRecordIndex = null;
+        this._currentRecord = null;
+        this._records = new Array();
+        this._bof = false;
+        this._eof = false;
+
+        let dataset = await this._datasource.load(fields, filter, maxRows);
+        for (let row of dataset) {
+            this._records.push(new DaoRecord(this._datasource.fieldDefs, row));
+        }
+        
+        if (this._records.length > 0) {
+            this._currentRecordIndex = 0;
+            this._currentRecord = this._records[this._currentRecordIndex];
+            this._bof = true;
+            this._eof = this._currentRecordIndex == this._records.length - 1;
+        }
+
+        return this.recordCount;
+    }
+
+    async save(): Promise<number> {
+        let recordsToUpdate = this._records.filter(r => r.hasChanged);
+        let recordsToInsert = this._records.filter(r => r.isNew);
+        return this._datasource.save(recordsToUpdate, recordsToInsert);
+    }
+
+    discard(): number {
+        let discardedCount = 0;
+        let changedRecords = this._records.filter(r => r.hasChanged);
+        for (let record of changedRecords) {
+            record.discard();
+            discardedCount++;
+        }
+
+        return discardedCount;
+    }
+
+    addRecord(): IDaoRecord;
+    addRecord(values?: Map<string, any>): IDaoRecord {
+        let row: any = [];
+        if (values) {
+            for (let fieldDef of this._datasource.fieldDefs) {
+                if (values.has(fieldDef.name)) {
+                    row.push(values.get(fieldDef.name));
+                }
+            }
+        }
+
+        let record = new DaoRecord(this._datasource.fieldDefs, row, true);
+        this._records.push(record);
+        return record;
+    }
+    
+    prev(): IDaoRecord {
+        if (this._currentRecordIndex == null) {
+            return null;
+        }
+        if (this._bof) {
+            return this._currentRecord;
+        }
+
+        if (this._currentRecordIndex > 0) {
+            this._currentRecordIndex--;
+            this._currentRecord = this._records[this._currentRecordIndex];
+        }
+        this._bof = this._currentRecordIndex == 0;
+        this._eof = this._currentRecordIndex == this._records.length - 1;
+
+        return this._currentRecord;
+    }
+
+    first(): IDaoRecord {
+        if (this._currentRecordIndex == null) {
+            return null;
+        }
+        if (this._bof) {
+            return this._currentRecord;
+        }
+
+        this._currentRecordIndex = 0;
+        this._currentRecord = this._records[this._currentRecordIndex];
+        this._bof = true;
+        this._eof = this._currentRecordIndex == this._records.length - 1;
+
+        return this._currentRecord;
+    }
+
+    next(): IDaoRecord {
+        if (this._currentRecordIndex == null) {
+            return null;
+        }
+        if (!this._eof) {
+            if (this._currentRecordIndex < this._records.length - 1) {
+                this._currentRecordIndex++;
+                this._currentRecord = this._records[this._currentRecordIndex];
+            }
+            this._bof = this._currentRecordIndex == 0;
+            if (this._currentRecordIndex == this._records.length - 1)
+            {
+                this._eof = true;
+            }
+        }
+
+        return this._currentRecord;
+    }
+
+    last(): IDaoRecord {
+        if (this._currentRecordIndex == null) {
+            return null;
+        }
+        if (this._eof) {
+            return this._currentRecord;
+        }
+
+        this._currentRecordIndex = this._records.length - 1;
+        this._currentRecord = this._records[this._currentRecordIndex];
+        this._bof = this._currentRecordIndex == 0;
+        this._eof = true;
+
+        return this._currentRecord;
+    }
 }
