+++ src/DAO/IDaoFactory.ts
@@ -1,8 +1,9 @@
-import { DaoTypes } from "./types";
+import { DaoType, DatasourceType } from "./types";
 import IDao from "./IDao";
 import IDaoFieldDef from "./IDaoFieldDef";
 
 export default interface IDaoFactory {
-    fieldDefs: Map<DaoTypes, Array<IDaoFieldDef>>;
-    getDao(objectName: string): Promise<IDao>;
+    //TypeScript won't allow static members.  Pointless.
+    //static fieldDefs: Map<DaoType, Array<IDaoFieldDef>>;
+    //static getDao(sourceType: DatasourceType, objectName: DaoType): Promise<IDao>;
 }
